requirements.txt
s3fs
zarr
python-dotenv
boto3
